+++ CHANGELOG.md
@@ -7,9 +7,11 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED
 - Added Marigold.
 - The Oasis's foliage and grass colors are no longer as bright.
 - Sunflower Plains have been revamped again, and are no longer an eyesore.
+- Renamed the Small Sponge to Sponge Bud.
 
 # Bug Fixes
 - Fixed a few tags that were still unfinished and excluded the Palm set.
 - Fixed the Carnation's dye recipe being in the `orange_dye` group.
 - Cherry particles are now much more responsive to wind, and no longer get removed upon spawning.
 - Wilder Wild's custom Sign types are now placed after the Mangrove Hanging Sign in the creative inventory, instead of between the Mangrove Sign and Mangrove Hanging Sign.
+- Ostrich Eggs are now placed before Sniffer Eggs in the creative inventory.